Fixed issue with TARGET variable.

+++ runCbt.java
@@ -17,9 +17,14 @@ import java.nio.file.*;
 import com.amazonaws.services.lambda.runtime.Context;
 
 public class runCbt {
+    public static void main(String[] args) {
+        runCbt run = new runCbt();
+        run.cbtHandler("usage", null);
+    }
+
     @SuppressWarnings("unchecked")
     public String cbtHandler(String args, Context Context) {
-        File dir = new File(".");
+        File dir = new File("."); //to test locally change this to "./cbt"
         File[] filesList = dir.listFiles();
         for (File file : filesList) {
             File src = file;
@@ -32,6 +37,9 @@ public class runCbt {
             }
         }
 
+        File start = new File("/tmp/startCbt.class");
+        String[] splitArgs = args.split("\\s+");
+        Object[] params = { splitArgs };
         File file = new File("/tmp");
 
         try {
@@ -40,7 +48,7 @@ public class runCbt {
             URL[] urls = new URL[]{url};
             ClassLoader cl = new URLClassLoader(urls);
             Class<?> cls = cl.loadClass("startCbt");
-            cls.getMethod("main", String[].class).invoke((Object) null, (Object) null);
+            cls.getMethod("main", String[].class).invoke((Object) null, params);
         } catch (Exception e) {
             e.printStackTrace();
             return "Failed.";
@@ -50,26 +58,28 @@ public class runCbt {
     }
 
     public static void copy(File source, File destination) throws IOException {
-        if (source.isDirectory()) {
-            destination.mkdir();
-            for (File file: source.listFiles()) {
-                copy(file, new File(destination.toString() + "/" + file.getName()));
+        if (!destination.exists()) {
+            if (source.isDirectory()) {
+                destination.mkdir();
+                for (File file: source.listFiles()) {
+                    copy(file, new File(destination.toString() + "/" + file.getName()));
+                }
+            } else {
+                InputStream inStream = new FileInputStream(source);
+                OutputStream outStream = new FileOutputStream(destination);
+                
+                byte[] buffer = new byte[1024];
+              
+                int length;
+                //copy the file content in bytes 
+                while ((length = inStream.read(buffer)) > 0){
+              
+                  outStream.write(buffer, 0, length);
+             
+                } 
+                inStream.close();
+                outStream.close();
             }
-        } else {
-            InputStream inStream = new FileInputStream(source);
-            OutputStream outStream = new FileOutputStream(destination);
-            
-            byte[] buffer = new byte[1024];
-          
-            int length;
-            //copy the file content in bytes 
-            while ((length = inStream.read(buffer)) > 0){
-          
-              outStream.write(buffer, 0, length);
-         
-            } 
-            inStream.close();
-            outStream.close();
         }
     }
 }
